+++ jestSetup.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line n/no-unpublished-require
 require('jest-fetch-mock').enableMocks()