+++ packages/runtime/src/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import { join, relative } from 'path'
 
 import type { NetlifyPlugin } from '@netlify/build'
@@ -80,12 +79,11 @@ const plugin: NetlifyPlugin = {
 
     checkNextSiteHasBuilt({ publish, failBuild })
 
-    const { appDir, basePath, i18n, images, target, ignore, trailingSlash, outdir, experimental } = await getNextConfig(
-      {
+    const { appDir, basePath, i18n, images, target, ignore, trailingSlash, outdir, experimental, routesManifest } =
+      await getNextConfig({
         publish,
         failBuild,
-      },
-    )
+      })
     await cleanupEdgeFunctions(constants)
 
     const middlewareManifest = await loadMiddlewareManifest(netlifyConfig)
@@ -155,7 +153,7 @@ const plugin: NetlifyPlugin = {
     const apiRoutes = await getExtendedApiRouteConfigs(publish, appDir)
 
     await generateFunctions(constants, appDir, apiRoutes)
-    await generatePagesResolver({ target, constants })
+    await generatePagesResolver(constants)
 
     await movePublicFiles({ appDir, outdir, publish })
 
@@ -186,7 +184,7 @@ const plugin: NetlifyPlugin = {
       apiRoutes,
     })
 
-    await writeEdgeFunctions(netlifyConfig)
+    await writeEdgeFunctions({ netlifyConfig, routesManifest })
   },
 
   async onPostBuild({
@@ -221,13 +219,18 @@ const plugin: NetlifyPlugin = {
     await checkZipSize(join(FUNCTIONS_DIST, `${ODB_FUNCTION_NAME}.zip`))
     const nextConfig = await getNextConfig({ publish, failBuild })
 
-    const { basePath, appDir } = nextConfig
+    const { basePath, appDir, experimental } = nextConfig
 
     generateCustomHeaders(nextConfig, headers)
 
     warnForProblematicUserRewrites({ basePath, redirects })
     warnForRootRedirects({ appDir })
     await warnOnApiRoutes({ FUNCTIONS_DIST })
+    if (experimental?.appDir) {
+      console.log(
+        '🧪 Thank you for testing "appDir" support on Netlify. For known issues and to give feedback, visit https://ntl.fyi/next-13-feedback',
+      )
+    }
   },
 }
 // The types haven't been updated yet
@@ -254,5 +257,3 @@ const nextRuntime = (
 }
 
 module.exports = nextRuntime
-
-/* eslint-enable max-lines */